refactor: modularize JS
