+++ web/themes/custom/cloudy/.browserslistrc
@@ -1,0 +1,9 @@
+# https://github.com/browserslist/browserslist
+# This list informs how Babel transforms our JS and Autoprefixer transforms our CSS
+# To see what browsers this config results in, run `npx browserslist`
+
+> 5% in US
+last 2 major versions
+ie 11
+not ie < 11
+not ie_mob >= 1
