+++ web/themes/custom/cloudy/babel.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "useBuiltIns": false,
+      }
+    ]
+  ]
+};
