+++ web/themes/custom/cloudy/cloudy.info.yml
@@ -6,7 +6,8 @@ core: 8.x
 base theme: classy
 
 libraries:
-  - cloudy/global-styling
+  - cloudy/bootstrap
+  - cloudy/global
 
 stylesheets-remove:
   - '@extlink/extlink.css'
