+++ web/themes/custom/cloudy/cloudy.libraries.yml
@@ -1,9 +1,14 @@
-global-styling:
+bootstrap:
+  # Note: the Bootstrap CSS is currently contained within the `global` library
   js:
-    dist/main.bundle.js: {}
+    dist/bootstrap.bundle.js: {}
+
+global:
+  js:
+    dist/cloudy.bundle.js: {}
   css:
     component:
-      dist/style.bundle.css: {}
+      dist/cloudy.css: {}
       https://use.fontawesome.com/releases/v5.2.0/css/all.css:
         type: external
         minified: true
