+++ web/themes/custom/cloudy/package.json
@@ -7,7 +7,7 @@
     "build": "npm-run-all --parallel --print-label build:*",
     "build:pl": "patternlab build --config patternlab-config.json",
     "build:pl:full": "npm run clean:pl && npm run build:pl",
-    "build:wp": "webpack --mode production --no-progress",
+    "build:wp": "NODE_ENV=production webpack --no-progress",
     "clean": "npm-run-all --parallel --print-label clean:*",
     "clean:pl": "rm -rf ../../../pattern-lab",
     "clean:wp": "rm -rf dist",
@@ -15,8 +15,8 @@
     "start": "npm-run-all --parallel --print-label watch:*",
     "test": "echo \"Error: no test specified\" && exit 1",
     "watch": "npm-run-all --parallel --print-label watch:*",
-    "watch:pl": "nodemon --ext twig,json --watch pattern-lab --watch src --exec 'npm run build:pl' ",
-    "watch:wp": "webpack-cli --config ./webpack.config.js --mode development --watch --hide-modules --display minimal"
+    "watch:pl": "nodemon --ext twig,json --watch pattern-lab --watch src --exec 'npm run build:pl:full' ",
+    "watch:wp": "webpack-cli --config ./webpack.config.js --watch --hide-modules --display minimal"
   },
   "keywords": [
     "SASS",
@@ -34,7 +34,8 @@
     "bootstrap4-chosen": "^1.0.0-alpha.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.4",
+    "@babel/core": "^7.7.5",
+    "@babel/preset-env": "^7.7.6",
     "@fullhuman/postcss-purgecss": "^1.2.0",
     "autoprefixer": "^9.7.3",
     "babel-loader": "^8.0.6",
@@ -49,7 +50,7 @@
     "jquery": "^3.4.1",
     "mini-css-extract-plugin": "^0.5.0",
     "node-sass": "^4.13.0",
-    "nodemon": "^2.0.1",
+    "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "popper.js": "^1.16.0",
     "postcss": "^6.0.19",
