--- web/themes/custom/cloudy/pattern-lab/_layouts/.gitkeep
@@ -1,1 +1,0 @@
-keeping this directory
