+++ web/themes/custom/cloudy/pattern-lab/_meta/_00-head.twig
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width" />
 
-    <link rel="stylesheet" href="/themes/custom/cloudy/dist/style.bundle.css?{{ cacheBuster }}" media="all" />
+    <link rel="stylesheet" href="/themes/custom/cloudy/dist/cloudy.css?{{ cacheBuster }}" media="all" />
     <link rel="stylesheet" href="/themes/custom/cloudy/pattern-lab/css/pattern-scaffolding.css?{{ cacheBuster }}" media="all" />
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css" media="all" />
