+++ web/themes/custom/cloudy/pattern-lab/_meta/_01-foot.twig
@@ -2,7 +2,10 @@
   <!--DO NOT REMOVE-->
   {{ patternLabFoot | raw }}
 
-  <script async src="/themes/custom/cloudy/dist/main.bundle.js" type="application/javascript"></script>
+{# START: Theme JS that will get loaded in Drupal as well #}
+<script src="/themes/custom/cloudy/dist/bootstrap.bundle.js"></script>
+<script src="/themes/custom/cloudy/dist/cloudy.bundle.js"></script>
+{# END: Theme JS #}
 
   </body>
 </html>
