+++ web/themes/custom/cloudy/src/cloudy.js
@@ -1,0 +1,4 @@
+import './js/libs/jquery-plugin-fitvids';
+import './components/alerts/alerts';
+import './components/dialog/dialog';
+import './components/tabs/tabs';
