+++ web/themes/custom/cloudy/src/cloudy.scss
@@ -2,13 +2,13 @@
 // ** Core (Non-printing, used by the preprocessor)
 // --------------
 //  Vars + Functions
-@import 'theme-variables';
-@import 'bootstrap-overrides';
+@import 'css/theme-variables';
+@import 'css/bootstrap-overrides';
 @import '~bootstrap/scss/functions';
 @import '~bootstrap/scss/variables';
 //  Mixins
 @import '~bootstrap/scss/mixins';
-@import 'mixins';
+@import 'css/mixins';
 
 
 // --------------
@@ -56,23 +56,26 @@
 @import '~bootstrap/scss/print';
 @import '~bootstrap4-chosen/bootstrap4-chosen.css';
 
-//  Elements
-//@todo create the below file and populate with all rules for bare html elements (i.e. detail, p, a, etc)
-//@todo There is some redundancy with the intention of 'elements' and the existing typography work, consider either importing typography through elements, breaking typography up, or leaving separated
-//@import 'elements';
-
 //  Typography
-@import 'typography';
+@import 'css/typography';
+
+//  Elements
+@import 'css/elements';
 
 // --------------
 // ** Components
 // --------------
-@import 'components';
+@import 'css/components';
+
+// --------------
+// ** Layouts
+// --------------
+@import 'css/layouts';
 
 // --------------
 // ** Utilities/Misc
 // --------------
-@import 'overrides';
+@import 'css/overrides';
 
 //@todo Remove all Tailwind-ism
 @tailwind utilities;
