+++ web/themes/custom/cloudy/src/components/tabs/tabs.scss
