+++ web/themes/custom/cloudy/src/css/_components.scss
@@ -1,6 +1,6 @@
 @import '../components/_address';
 @import '../components/_affix';
-@import '../components/_alerts';
+@import '../components/alerts/alerts';
 @import '../components/_badge';
 @import '../components/_blockquote';
 @import '../components/_breadcrumb';
@@ -9,7 +9,7 @@
 @import '../components/_columns';
 @import '../components/content';
 @import '../components/_council';
-@import '../components/dialog';
+@import '../components/dialog/dialog';
 @import '../components/dropbutton';
 @import '../components/_electeds';
 @import '../components/_entity-browser';
@@ -33,7 +33,7 @@
 @import '../components/_skip-link';
 @import '../components/_tables';
 @import '../components/_tablesort-indicator';
-@import '../components/_tabs';
+@import '../components/tabs/tabs';
 @import '../components/_textarea';
 @import '../components/_user';
 @import '../components/_vertical-tabs.component';
